requirements.txt
selenium==4.23.1
openpyxl==3.1.5
pytest==8.3.2
pytest-xdist==3.6.1
pytest-html==4.1.1
allure-pytest==2.13.5